Load test WoW data for view tests.
