+++ pytest.ini
@@ -1,5 +1,5 @@
 [pytest]
-DJANGO_SETTINGS_MODULE = project.settings
+DJANGO_SETTINGS_MODULE = project.settings_pytest
 norecursedirs = .git .venv .vscode node_modules venv
 addopts = --doctest-modules
 filterwarnings =
